requirements.txt
Flask
celery
redis
opencv-python
Pillow
numpy
requests
python-dotenv
openai